requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.7.0
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
addict=2.4.0=pypi_0
aenum=2.2.6=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.11.0=pypi_0
asttokens=3.0.1=pyhd8ed1ab_0
attr=2.5.2=h39aace5_0
attrs=25.4.0=pypi_0
bidict=0.23.1=pypi_0
blinker=1.9.0=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=hda65f42_8
ca-certificates=2025.11.12=hbd8a1cb_0
certifi=2025.11.12=pypi_0
charset-normalizer=3.4.4=pypi_0
click=8.3.0=pypi_0
comm=0.2.3=pyhe01879c_0
configargparse=1.7.1=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
dash=3.3.0=pypi_0
debugpy=1.8.17=py310h25320af_0
decorator=5.2.1=pyhd8ed1ab_0
et-xmlfile=2.0.0=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.1=pyhd8ed1ab_0
fastjsonschema=2.21.2=pypi_0
filelock=3.19.1=pypi_0
flask=2.0.1=pypi_0
flask-cors=3.0.10=pypi_0
flask-socketio=5.5.1=pypi_0
fline=0.1.3=pypi_0
fonttools=4.60.1=pypi_0
fsspec=2025.9.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
glm=0.4.4=pypi_0
grad-cam=1.5.5=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.2.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=1.1.4=pypi_0
icu=75.1=he02047a_0
idna=3.11=pypi_0
imageio=2.37.2=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
ipykernel=7.1.0=pyha191276_0
ipython=8.37.0=pyh8f84b5b_0
ipywidgets=8.1.8=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
joblib=1.5.2=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.9.1=pyhc90fa1f_0
jupyterlab-widgets=3.0.16=pypi_0
keyutils=1.6.3=hb9d3cd8_0
kiwisolver=1.4.9=pypi_0
krb5=1.21.3=h659f571_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.45=h1aa0949_0
libcap=2.77=h3ff7636_0
libedit=3.1.20250104=pl5321h7949ede_0
libevent=2.1.12=hf998b51_1
libexpat=2.7.1=hecca717_0
libfabric=2.3.1=ha770c72_1
libfabric1=2.3.1=h6c8fc0a_1
libffi=3.5.2=h9ec8514_0
libgcc=15.2.0=h767d61c_7
libgcc-ng=15.2.0=h69a702a_7
libgfortran=15.2.0=h69a702a_7
libgfortran5=15.2.0=hcd61629_7
libgomp=15.2.0=h767d61c_7
libhwloc=2.12.1=default_h7f8ec31_1002
libiconv=1.18=h3b78370_2
liblzma=5.8.1=hb9d3cd8_2
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hb9d3cd8_1
libpmix=5.0.8=h4bd6b51_2
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.51.0=hee844dc_0
libstdcxx=15.2.0=h8f9b012_7
libstdcxx-ng=15.2.0=h4852527_7
libsystemd0=257.10=hd0affe5_2
libudev1=257.10=hd0affe5_2
libuuid=2.41.2=he9a06e4_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.15.1=h26afc86_0
libxml2-16=2.15.1=ha9997c6_0
libzlib=1.3.1=hb9d3cd8_2
markupsafe=2.1.5=pypi_0
matplotlib=3.10.7=pypi_0
matplotlib-inline=0.2.1=pyhd8ed1ab_0
mpi=1.0.1=openmpi
mpmath=1.3.0=pypi_0
narwhals=2.12.0=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
noise=1.2.2=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-ml-py=12.560.30=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
open3d=0.19.0=pypi_0
opencv-python=4.12.0.88=pypi_0
openmpi=5.0.8=h2fe1745_108
openpyxl=3.1.5=pypi_0
openssl=3.6.0=h26f9b46_0
packaging=25.0=pyh29332c3_1
pandas=2.3.1=pypi_0
parso=0.8.5=pyhcf101f3_0
pegl=0.2a1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=pypi_0
pip=25.3=pyh8b19718_0
platformdirs=4.5.0=pyhcf101f3_0
plotly=6.4.0=pypi_0
prompt-toolkit=3.0.52=pyha770c72_0
protobuf=6.33.1=pypi_0
psutil=7.1.3=py310h139afa4_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pydantic=2.12.4=pypi_0
pydantic-core=2.41.5=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pynvml=11.0.0=pypi_0
pyparsing=3.2.5=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.10.19=h3c07f61_2_cpython
python-dateutil=2.9.0.post0=pyhe01879c_2
python-engineio=4.12.3=pypi_0
python-socketio=5.15.0=pypi_0
python_abi=3.10=8_cp310
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=py310h4f33d48_0
rdma-core=60.0=hecca717_0
readline=8.2=h8c095d6_2
referencing=0.37.0=pypi_0
requests=2.32.5=pypi_0
retrying=1.4.2=pypi_0
rpds-py=0.29.0=pypi_0
safetensors=0.6.2=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
sentry-sdk=2.44.0=pypi_0
setuptools=80.9.0=pyhff2d567_0
shellingham=1.5.4=pypi_0
simple-websocket=1.1.0=pypi_0
six=1.17.0=pyhe01879c_1
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.5.10=pypi_0
timm=1.0.22=pypi_0
tk=8.6.13=noxft_ha0e22de_103
torch=2.7.1+cu118=pypi_0
torchaudio=2.7.1+cu118=pypi_0
torchvision=0.22.1+cu118=pypi_0
tornado=6.5.2=py310h7c4b9e2_2
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
triton=3.3.1=pypi_0
ttach=0.0.3=pypi_0
typer-slim=0.20.0=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.2=pypi_0
ucc=1.5.1=hb729f83_0
ucx=1.19.0=h63b5c0b_5
urllib3=2.5.0=pypi_0
wandb=0.23.0=pypi_0
wcwidth=0.2.14=pyhd8ed1ab_0
webdataset=1.0.2=pypi_0
werkzeug=2.3.0=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
widgetsnbextension=4.0.15=pypi_0
wsproto=1.3.2=pypi_0
zeromq=4.3.5=h387f397_9
zipp=3.23.0=pyhcf101f3_1
zstd=1.5.7=hb8e6e7a_2
